# DBZ-3770 Optimize ObjectMapper initialization by replacing repeated initialization with one time initilization